refactor: move and re-export button component
Moved the button component to the 'action' directory for better organization and consistency with other action components. Updated the exports in 'index.tsx' to reflect the new structure.
